Qwerty Melior Support for Revolve Generator (#1)
* updates to qwerty_melior for ApplyRevolveGeneratorAttribute

* updating revolve attribute stuff

* formatting fix

* Remove unused import

I am having a vivid vision of me forgetting to remove this or missing it
in the PR.

* Add method for printing an op to a file

* removed fourier from qwerty_melior

* updates to qwerty_melior for ApplyRevolveGeneratorAttribute

* updating revolve attribute stuff

* formatting fix

* Remove unused import

I am having a vivid vision of me forgetting to remove this or missing it
in the PR.

* Add method for printing an op to a file

* removed fourier from qwerty_melior

* Add newtype for `io::Error`

Gets unit tests passing

* Disable tosa tablegen

This makes unit tests fail. But I have no idea what it even is and don't
care one bit. No offense to Mr. and Mrs. TOSA.

* Remove unnecessary Display impl

* PR fixes, minor changes

* i made merge mistakes

* missed a comment to delete

* i missed something in my merge mistakes

---------

Co-authored-by: Austin Adams <git@austinjadams.com>